requirements.txt
fastapi
uvicorn[standard]
ovh
aiosqlite
pyyaml
